Enable Async Eval
These checks are great for the serial case but break down when we do async eval; tags can end up concurrent and out-of-order. I did my best to remove the minimum number of checks to enable async eval. Would be happy to discuss how we can re-do these checks to be async friendly in the future.